refactor: split anti-slop into modules, remove style violations
- Split single 295-line file into focused modules:
  - anti-slop.ts (entry point, 49 lines)
  - config.ts (TOML parsing)
  - ruff.ts (Ruff subprocess wrapper)
  - types.ts (shared Violation interface)
  - rules/smart-punctuation.ts (SLOP001)
  - rules/except-pass.ts (SLOP002)

- Remove decorative banner comments (code-style violation)
- Remove paraphrasing docstring on parseTomlConfig
- Keep explanatory comments that document why (Python ast drops
  comments, Ruff exit codes, subprocess fallback)
