+++ .pi/agent/skills/anti-slop/types.ts
@@ -1,0 +1,7 @@
+export interface Violation {
+  file: string;
+  line: number;
+  col: number;
+  rule: string;
+  message: string;
+}
